--- modules/ballerina-core/src/main/java/org/ballerinalang/bre/MicroTransactionContext.java
@@ -1,36 +1,0 @@
-/*
- * Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.ballerinalang.bre;
-
-/**
- * {@code MicroTransactionContext} represents the context of the micro-transaction.
- */
-public class MicroTransactionContext {
-    private final double contextVersion;
-    private final String identifier;
-    private final String coordinationType;
-    private final String registerAt;
-
-    public MicroTransactionContext(String identifier, String coordinationType, double contextVersion, String registerAt) {
-        this.identifier = identifier;
-        this.coordinationType = coordinationType;
-        this.contextVersion = contextVersion;
-        this.registerAt = registerAt;
-    }
-}
